refactor(SI-2628): pr comment edit - moved AlertDialog from wm class
