+++ lib/features/places_list/screens/places_list_screen/places_list_screen_wm.dart
@@ -4,9 +4,9 @@ import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 import 'package:provider/provider.dart';
-import 'package:surf_study_project/assets/strings/app_strings.dart';
 import 'package:surf_study_project/features/app/di/app_scope.dart';
 import 'package:surf_study_project/features/common/domain/entity/place.dart';
+import 'package:surf_study_project/features/common/widgets/alert_dialog/location_alert_dialog.dart';
 import 'package:surf_study_project/features/geolocation/bloc/geolocation_bloc.dart';
 import 'package:surf_study_project/features/navigation/domain/entity/app_route_paths.dart';
 import 'package:surf_study_project/features/navigation/service/router.dart';
@@ -137,7 +137,7 @@ class PlacesListScreenWidgetModel
       context: context,
       barrierDismissible: false,
       builder: (context) {
-        return _AlertDialog(
+        return LocationAlertDialog(
           onAlertCloseTap: _onAlertCloseTap,
           onAlertSettingsTap: _onAlertSettingsTap,
         );
@@ -162,37 +162,4 @@ abstract class IPlacesListScreenWidgetModel extends IWidgetModel {
 
   /// refresh list
   Future<void> onRefresh();
-}
-
-/// [AlertDialog] showing message to permit location
-class _AlertDialog extends StatelessWidget {
-  final VoidCallback onAlertCloseTap;
-  final VoidCallback onAlertSettingsTap;
-
-  const _AlertDialog({
-    required this.onAlertCloseTap,
-    required this.onAlertSettingsTap,
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: const Text(AppStrings.filterAlertTitle),
-      content: const SingleChildScrollView(
-        child: Text(AppStrings.filterAlertMessage),
-      ),
-      actions: <Widget>[
-        TextButton(
-          child: const Text(AppStrings.filterAlertClose),
-          onPressed: onAlertCloseTap,
-          // onPressed: router.pop,
-        ),
-        TextButton(
-          child: const Text(AppStrings.filterAlertGotoSettings),
-          onPressed: onAlertSettingsTap,
-        ),
-      ],
-    );
-  }
 }
